+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.40.0 - TBD
+
+#### Enhancements
+- Added `adjustment_factors.get_range(...)` method for `Reference` client
+- Added `security_master.get_range(...)` method for `Reference` client
+- Added `security_master.get_last(...)` method for `Reference` client
+
 ## 0.39.3 - 2024-08-20
 
 #### Enhancements